Join behavior change + for each module split
