+++ src/operations/mod.rs
@@ -2,12 +2,24 @@ use crate::{
     relation::{CheckRelations, EdgeIter, EdgeWQ, EdgeWQItem, IterRelations, Relation, ZstOrPanic},
     tuple_traits::*,
 };
+
 use bevy::ecs::{
     entity::Entity,
     query::{ReadOnlyWorldQuery, WorldQuery},
     system::Query,
 };
-use std::{borrow::Borrow, collections::VecDeque, marker::PhantomData};
+
+use std::{borrow::Borrow, marker::PhantomData};
+
+mod for_each;
+mod for_each_3arity;
+mod for_each_4arity;
+mod for_each_5arity;
+
+pub use for_each::*;
+pub use for_each_3arity::*;
+pub use for_each_4arity::*;
+pub use for_each_5arity::*;
 
 /// Struct to track inner product iteration.
 pub struct EdgeProduct<'a, const N: usize> {
@@ -863,8 +875,8 @@ where
 /// E6:::pink --R--> E2:::pink
 /// E7:::pink --R--> E2:::pink
 ///
-/// E8:::pink --R--> E3
-/// E9:::pink --R--> E3
+/// E8 --R--> E3
+/// E9 --R--> E3
 /// ```
 pub enum ControlFlow {
     /// Continue to next permutation.
@@ -893,585 +905,6 @@ pub enum ControlFlow {
 impl From<()> for ControlFlow {
     fn from(_: ()) -> Self {
         ControlFlow::Continue
-    }
-}
-
-/// A trait to iterate relation queries.
-///
-/// When iterating:
-/// - Diamonds
-/// - Cycles
-/// - Joins where multiple entities have the same target
-///
-/// References to the same entities components can be produced more than once which is why this
-/// problem cannot be solved with `unsafe`. It requires lending iterators which are not expressable
-/// with current GATs so to work around this "lifetime trapping" is used instead.
-/// The closure parameters cannot escape the closure.
-///
-/// For any control query `Query<(X, Relations<..>)>`:
-/// - If there is only joins: Permutations of valid entities from the joined queries will be looped
-/// through. The left parameter will be the `X` item and the right parameter will be a tuple of all
-/// the query fetch parameters from the joined queries.
-/// - If there is only hierarchy traversals: Traversable ancestor-descendant permutations that
-/// belong to the control query will be looped through. For descents the left parameter will be the
-/// `X` item of an ancestor and the right parameter will be the `X` item of an immediate descendant.
-/// For ascents this is the other way around.
-///
-/// See [`ControlFlow`] for control flow options and [`ForEachPermutations3Arity`] for the loop
-/// behavior of operations with hierarchy traversals and 1 or more join.
-pub trait ForEachPermutations<const N: usize> {
-    type Left<'l>;
-    type Right<'r>;
-
-    fn for_each<Func, Ret>(self, func: Func)
-    where
-        Ret: Into<ControlFlow>,
-        Func: for<'f, 'l, 'r> FnMut(&'f mut Self::Left<'l>, Self::Right<'r>) -> Ret;
-}
-
-impl<Q, R, F, JoinedTypes, JoinedQueries, const N: usize> ForEachPermutations<N>
-    for Operations<&'_ Query<'_, '_, (Q, Relations<R>), F>, JoinedTypes, JoinedQueries, ()>
-where
-    Q: WorldQuery,
-    R: RelationSet,
-    F: ReadOnlyWorldQuery,
-    JoinedTypes: Product<N>,
-    JoinedQueries: for<'a> Joinable<'a, N>,
-{
-    type Left<'l> = <<Q as WorldQuery>::ReadOnly as WorldQuery>::Item<'l>;
-    type Right<'r> = <JoinedQueries as Joinable<'r, N>>::Out;
-
-    fn for_each<Func, Ret>(mut self, mut func: Func)
-    where
-        Ret: Into<ControlFlow>,
-        Func: for<'f, 'l, 'r> FnMut(&'f mut Self::Left<'l>, Self::Right<'r>) -> Ret,
-    {
-        for (mut control, relations) in self.control.iter() {
-            let mut edge_product = JoinedTypes::product(relations.edges);
-            let mut matches = [false; N];
-
-            while let Some(entities) = edge_product.advance(matches) {
-                matches = Joinable::check(&self.joined_queries, entities);
-
-                if matches.iter().any(|b| !b) {
-                    continue;
-                }
-
-                match func(
-                    &mut control,
-                    Joinable::join(&mut self.joined_queries, entities),
-                )
-                .into()
-                {
-                    ControlFlow::Continue => {}
-                    ControlFlow::Exit | ControlFlow::Conclude => return,
-                    ControlFlow::Walk | ControlFlow::Probe => break,
-                    ControlFlow::FastForward(n) if n < N => {
-                        matches[n] = false;
-                    }
-                    _ => {}
-                }
-            }
-        }
-    }
-}
-
-impl<Q, R, F, JoinedTypes, JoinedQueries, const N: usize> ForEachPermutations<N>
-    for Operations<&'_ mut Query<'_, '_, (Q, Relations<R>), F>, JoinedTypes, JoinedQueries, ()>
-where
-    Q: WorldQuery,
-    R: RelationSet,
-    F: ReadOnlyWorldQuery,
-    JoinedTypes: Product<N>,
-    JoinedQueries: for<'a> Joinable<'a, N>,
-{
-    type Left<'l> = <Q as WorldQuery>::Item<'l>;
-    type Right<'r> = <JoinedQueries as Joinable<'r, N>>::Out;
-
-    fn for_each<Func, Ret>(mut self, mut func: Func)
-    where
-        Ret: Into<ControlFlow>,
-        Func: for<'f, 'l, 'r> FnMut(&'f mut Self::Left<'l>, Self::Right<'r>) -> Ret,
-    {
-        for (mut control, relations) in self.control.iter_mut() {
-            let mut edge_product = JoinedTypes::product(relations.edges);
-            let mut matches = [false; N];
-
-            while let Some(entities) = edge_product.advance(matches) {
-                matches = Joinable::check(&self.joined_queries, entities);
-
-                if matches.iter().any(|b| !b) {
-                    continue;
-                }
-
-                match func(
-                    &mut control,
-                    Joinable::join(&mut self.joined_queries, entities),
-                )
-                .into()
-                {
-                    ControlFlow::Continue => {}
-                    ControlFlow::Exit | ControlFlow::Conclude => return,
-                    ControlFlow::Walk | ControlFlow::Probe => break,
-                    ControlFlow::FastForward(n) if n < N => {
-                        matches[n] = false;
-                    }
-                    _ => {}
-                }
-            }
-        }
-    }
-}
-
-impl<Q, R, F, T, E, I> ForEachPermutations<0>
-    for Operations<&'_ Query<'_, '_, (Q, Relations<R>), F>, (), (), T, I>
-where
-    Q: WorldQuery,
-    R: RelationSet,
-    F: ReadOnlyWorldQuery,
-    T: EdgeQuery,
-    E: Borrow<Entity>,
-    I: IntoIterator<Item = E>,
-{
-    type Left<'l> = <<Q as WorldQuery>::ReadOnly as WorldQuery>::Item<'l>;
-    type Right<'r> = <<Q as WorldQuery>::ReadOnly as WorldQuery>::Item<'r>;
-
-    fn for_each<Func, Ret>(self, mut func: Func)
-    where
-        Ret: Into<ControlFlow>,
-        Func: for<'f, 'l, 'r> FnMut(&'f mut Self::Left<'l>, Self::Right<'r>) -> Ret,
-    {
-        let mut queue = self
-            .starts
-            .into_iter()
-            .map(|e| *e.borrow())
-            .collect::<VecDeque<Entity>>();
-
-        'queue: while let Some(entity) = queue.pop_front() {
-            let Ok((mut control, relations)) = self.control.get(entity) else {
-                continue;
-            };
-
-            for e in T::entities(&relations.edges) {
-                let Ok(traversal_item) = self.control.get(e) else {
-                    continue;
-                };
-
-                match func(&mut control, traversal_item.0).into() {
-                    ControlFlow::Exit => return,
-                    ControlFlow::Conclude => {
-                        continue 'queue;
-                    }
-                    ControlFlow::Probe => {
-                        queue.clear();
-                        queue.push_back(e);
-                        continue 'queue;
-                    }
-                    _ => {}
-                }
-            }
-
-            queue.extend(T::entities(&relations.edges));
-        }
-    }
-}
-
-impl<Q, R, F, T, E, I> ForEachPermutations<0>
-    for Operations<&'_ mut Query<'_, '_, (Q, Relations<R>), F>, (), (), T, I>
-where
-    Q: WorldQuery,
-    R: RelationSet,
-    F: ReadOnlyWorldQuery,
-    T: EdgeQuery,
-    E: Borrow<Entity>,
-    I: IntoIterator<Item = E>,
-{
-    type Left<'l> = <Q as WorldQuery>::Item<'l>;
-    type Right<'r> = <Q as WorldQuery>::Item<'r>;
-
-    fn for_each<Func, Ret>(self, mut func: Func)
-    where
-        Ret: Into<ControlFlow>,
-        Func: for<'f, 'l, 'r> FnMut(&'f mut Self::Left<'l>, Self::Right<'r>) -> Ret,
-    {
-        let mut queue = self
-            .starts
-            .into_iter()
-            .map(|e| *e.borrow())
-            .collect::<VecDeque<Entity>>();
-
-        'queue: while let Some(entity) = queue.pop_front() {
-            // SAFETY: Self referential relations are impossible so this is always safe.
-            let Ok((mut control, relations)) = (unsafe { self.control.get_unchecked(entity) })
-            else {
-                continue;
-            };
-
-            for e in T::entities(&relations.edges) {
-                // SAFETY: Self referential relations are impossible so this is always safe.
-                let Ok(traversal_item) = (unsafe { self.control.get_unchecked(e) }) else {
-                    continue;
-                };
-
-                match func(&mut control, traversal_item.0).into() {
-                    ControlFlow::Exit => return,
-                    ControlFlow::Conclude => {
-                        continue 'queue;
-                    }
-                    ControlFlow::Probe => {
-                        queue.clear();
-                        queue.push_back(e);
-                        continue 'queue;
-                    }
-                    _ => {}
-                }
-            }
-
-            queue.extend(T::entities(&relations.edges));
-        }
-    }
-}
-
-impl<Q, R, F, T, E, I, Acc, Err, Init, Fold> ForEachPermutations<0>
-    for Operations<&'_ Query<'_, '_, (Q, Relations<R>), F>, (), (), T, I, Init, Fold>
-where
-    Q: WorldQuery,
-    R: RelationSet,
-    F: ReadOnlyWorldQuery,
-    T: EdgeQuery,
-    E: Borrow<Entity>,
-    I: IntoIterator<Item = E>,
-    Init: for<'a> FnMut(&mut <<Q as WorldQuery>::ReadOnly as WorldQuery>::Item<'a>) -> Acc,
-    Fold: for<'a> FnMut(
-        Acc,
-        <<Q as WorldQuery>::ReadOnly as WorldQuery>::Item<'a>,
-    ) -> Result<Acc, Err>,
-{
-    type Left<'l> = (
-        <<Q as WorldQuery>::ReadOnly as WorldQuery>::Item<'l>,
-        Result<Acc, Err>,
-    );
-    type Right<'r> = <<Q as WorldQuery>::ReadOnly as WorldQuery>::Item<'r>;
-
-    fn for_each<Func, Ret>(mut self, mut func: Func)
-    where
-        Ret: Into<ControlFlow>,
-        Func: for<'f, 'l, 'r> FnMut(&'f mut Self::Left<'l>, Self::Right<'r>) -> Ret,
-    {
-        let mut queue = self
-            .starts
-            .into_iter()
-            .map(|e| *e.borrow())
-            .collect::<VecDeque<Entity>>();
-
-        'queue: while let Some(entity) = queue.pop_front() {
-            let Ok((mut control, relations)) = self.control.get(entity) else {
-                continue;
-            };
-
-            let mut acc = Ok::<_, Err>((self.init)(&mut control));
-
-            for e in T::entities(&relations.edges) {
-                let Ok(traversal_item) = self.control.get(e) else {
-                    continue;
-                };
-
-                let Ok(acc_ok) = acc else {
-                    break;
-                };
-
-                acc = (self.fold)(acc_ok, traversal_item.0);
-            }
-
-            let mut left = (control, acc);
-
-            for e in T::entities(&relations.edges) {
-                let Ok(traversal_item) = self.control.get(e) else {
-                    continue;
-                };
-
-                match func(&mut left, traversal_item.0).into() {
-                    ControlFlow::Exit => return,
-                    ControlFlow::Conclude => {
-                        continue 'queue;
-                    }
-                    ControlFlow::Probe => {
-                        queue.clear();
-                        queue.push_back(e);
-                        continue 'queue;
-                    }
-                    _ => {}
-                }
-            }
-
-            queue.extend(T::entities(&relations.edges));
-        }
-    }
-}
-
-impl<Q, R, F, T, E, I, Acc, Err, Init, Fold> ForEachPermutations<0>
-    for Operations<&'_ mut Query<'_, '_, (Q, Relations<R>), F>, (), (), T, I, Init, Fold>
-where
-    Q: WorldQuery,
-    R: RelationSet,
-    F: ReadOnlyWorldQuery,
-    T: EdgeQuery,
-    E: Borrow<Entity>,
-    I: IntoIterator<Item = E>,
-    Init: for<'a> FnMut(&mut <Q as WorldQuery>::Item<'a>) -> Acc,
-    Fold: for<'a> FnMut(Acc, <Q as WorldQuery>::Item<'a>) -> Result<Acc, Err>,
-{
-    type Left<'l> = (<Q as WorldQuery>::Item<'l>, Result<Acc, Err>);
-    type Right<'r> = <Q as WorldQuery>::Item<'r>;
-
-    fn for_each<Func, Ret>(mut self, mut func: Func)
-    where
-        Ret: Into<ControlFlow>,
-        Func: for<'f, 'l, 'r> FnMut(&'f mut Self::Left<'l>, Self::Right<'r>) -> Ret,
-    {
-        let mut queue = self
-            .starts
-            .into_iter()
-            .map(|e| *e.borrow())
-            .collect::<VecDeque<Entity>>();
-
-        'queue: while let Some(entity) = queue.pop_front() {
-            // SAFETY: Self referential relations are impossible so this is always safe.
-            let Ok((mut control, relations)) = (unsafe { self.control.get_unchecked(entity) })
-            else {
-                continue;
-            };
-
-            let mut acc = Ok::<_, Err>((self.init)(&mut control));
-
-            for e in T::entities(&relations.edges) {
-                // SAFETY: Self referential relations are impossible so this is always safe.
-                let Ok(traversal_item) = (unsafe { self.control.get_unchecked(e) }) else {
-                    continue;
-                };
-
-                let Ok(acc_ok) = acc else {
-                    break;
-                };
-
-                acc = (self.fold)(acc_ok, traversal_item.0);
-            }
-
-            let mut left = (control, acc);
-
-            for e in T::entities(&relations.edges) {
-                // SAFETY: Self referential relations are impossible so this is always safe.
-                let Ok(traversal_item) = (unsafe { self.control.get_unchecked(e) }) else {
-                    continue;
-                };
-
-                match func(&mut left, traversal_item.0).into() {
-                    ControlFlow::Exit => return,
-                    ControlFlow::Conclude => {
-                        continue 'queue;
-                    }
-                    ControlFlow::Probe => {
-                        queue.clear();
-                        queue.push_back(e);
-                        continue 'queue;
-                    }
-                    _ => {}
-                }
-            }
-
-            queue.extend(T::entities(&relations.edges));
-        }
-    }
-}
-
-/// A 3 arity version of [`ForEachPermutations`] for when operations feature a traversal with 1 or
-/// more joins. Will iterate through hierarchy permutations and join permutations together.
-/// - The left and middle paramaters will be an ancestor/descendant pairs.
-/// - The rightmost parameter will be a tuple of all the query fetch parameters from joined queries
-/// where the entity being joined on is the same entity that is the middle parameter. This is the
-/// ancestor or descendant depending on if the traversal is an ascent or descent.
-pub trait ForEachPermutations3Arity<const N: usize> {
-    type Left<'l>;
-    type Middle<'m>;
-    type Right<'r>;
-
-    fn for_each<Func, Ret>(self, func: Func)
-    where
-        Ret: Into<ControlFlow>,
-        Func: for<'f, 'l, 'm, 'r> FnMut(
-            &'f mut Self::Left<'l>,
-            &'f mut Self::Middle<'m>,
-            Self::Right<'r>,
-        ) -> Ret;
-}
-
-impl<Q, R, F, T, E, I, JoinedTypes, JoinedQueries, const N: usize> ForEachPermutations3Arity<N>
-    for Operations<&'_ Query<'_, '_, (Q, Relations<R>), F>, JoinedTypes, JoinedQueries, T, I>
-where
-    Q: WorldQuery,
-    R: RelationSet,
-    F: ReadOnlyWorldQuery,
-    T: EdgeQuery,
-    E: Borrow<Entity>,
-    I: IntoIterator<Item = E>,
-    JoinedTypes: Product<N>,
-    JoinedQueries: for<'a> Joinable<'a, N>,
-{
-    type Left<'l> = <<Q as WorldQuery>::ReadOnly as WorldQuery>::Item<'l>;
-    type Middle<'m> = <<Q as WorldQuery>::ReadOnly as WorldQuery>::Item<'m>;
-    type Right<'r> = <JoinedQueries as Joinable<'r, N>>::Out;
-
-    fn for_each<Func, Ret>(mut self, mut func: Func)
-    where
-        Ret: Into<ControlFlow>,
-        Func: for<'f, 'l, 'm, 'r> FnMut(
-            &'f mut Self::Left<'l>,
-            &'f mut Self::Middle<'m>,
-            Self::Right<'r>,
-        ) -> Ret,
-    {
-        let mut queue = self
-            .starts
-            .into_iter()
-            .map(|e| *e.borrow())
-            .collect::<VecDeque<Entity>>();
-
-        'queue: while let Some(entity) = queue.pop_front() {
-            let Ok((mut left_components, left_edges)) = self.control.get(entity) else {
-                continue;
-            };
-
-            for mid in T::entities(&left_edges.edges) {
-                let Ok((mut mid_components, mid_edges)) = self.control.get(mid) else {
-                    continue;
-                };
-
-                let mut edge_product = JoinedTypes::product(mid_edges.edges);
-                let mut matches = [false; N];
-
-                while let Some(entities) = edge_product.advance(matches) {
-                    matches = Joinable::check(&self.joined_queries, entities);
-
-                    if matches.iter().any(|b| !b) {
-                        continue;
-                    }
-
-                    match func(
-                        &mut left_components,
-                        &mut mid_components,
-                        Joinable::join(&mut self.joined_queries, entities),
-                    )
-                    .into()
-                    {
-                        ControlFlow::Continue => {}
-                        ControlFlow::Exit => return,
-                        ControlFlow::Walk => break,
-                        ControlFlow::FastForward(n) if n < N => {
-                            matches[n] = false;
-                        }
-                        ControlFlow::Conclude => {
-                            continue 'queue;
-                        }
-                        ControlFlow::Probe => {
-                            queue.clear();
-                            queue.push_back(mid);
-                            continue 'queue;
-                        }
-                        _ => {}
-                    }
-                }
-            }
-
-            queue.extend(T::entities(&left_edges.edges));
-        }
-    }
-}
-
-impl<Q, R, F, T, E, I, JoinedTypes, JoinedQueries, const N: usize> ForEachPermutations3Arity<N>
-    for Operations<&'_ mut Query<'_, '_, (Q, Relations<R>), F>, JoinedTypes, JoinedQueries, T, I>
-where
-    Q: WorldQuery,
-    R: RelationSet,
-    F: ReadOnlyWorldQuery,
-    T: EdgeQuery,
-    E: Borrow<Entity>,
-    I: IntoIterator<Item = E>,
-    JoinedTypes: Product<N>,
-    JoinedQueries: for<'a> Joinable<'a, N>,
-{
-    type Left<'l> = <Q as WorldQuery>::Item<'l>;
-    type Middle<'m> = <Q as WorldQuery>::Item<'m>;
-    type Right<'r> = <JoinedQueries as Joinable<'r, N>>::Out;
-
-    fn for_each<Func, Ret>(mut self, mut func: Func)
-    where
-        Ret: Into<ControlFlow>,
-        Func: for<'f, 'l, 'm, 'r> FnMut(
-            &'f mut Self::Left<'l>,
-            &'f mut Self::Middle<'m>,
-            Self::Right<'r>,
-        ) -> Ret,
-    {
-        let mut queue = self
-            .starts
-            .into_iter()
-            .map(|e| *e.borrow())
-            .collect::<VecDeque<Entity>>();
-
-        'queue: while let Some(entity) = queue.pop_front() {
-            // SAFETY: Self referential relations are impossible so this is always safe.
-            let Ok((mut left_components, left_edges)) =
-                (unsafe { self.control.get_unchecked(entity) })
-            else {
-                continue;
-            };
-
-            for mid in T::entities(&left_edges.edges) {
-                // SAFETY: Self referential relations are impossible so this is always safe.
-                let Ok((mut mid_components, mid_edges)) =
-                    (unsafe { self.control.get_unchecked(mid) })
-                else {
-                    continue;
-                };
-
-                let mut edge_product = JoinedTypes::product(mid_edges.edges);
-                let mut matches = [false; N];
-
-                while let Some(entities) = edge_product.advance(matches) {
-                    matches = Joinable::check(&self.joined_queries, entities);
-
-                    if matches.iter().any(|b| !b) {
-                        continue;
-                    }
-
-                    match func(
-                        &mut left_components,
-                        &mut mid_components,
-                        Joinable::join(&mut self.joined_queries, entities),
-                    )
-                    .into()
-                    {
-                        ControlFlow::Continue => {}
-                        ControlFlow::Exit => return,
-                        ControlFlow::Walk => break,
-                        ControlFlow::FastForward(n) if n < N => {
-                            matches[n] = false;
-                        }
-                        ControlFlow::Conclude => {
-                            continue 'queue;
-                        }
-                        ControlFlow::Probe => {
-                            queue.clear();
-                            queue.push_back(mid);
-                            continue 'queue;
-                        }
-                        _ => {}
-                    }
-                }
-            }
-
-            queue.extend(T::entities(&left_edges.edges));
-        }
     }
 }
